% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill3DVectorImage.r
\name{displacementGrid2Transform}
\alias{displacementGrid2Transform}
\title{convert an object of class DisplacementGrid to deformation field}
\usage{
displacementGrid2Transform(x)
}
\arguments{
\item{x}{object of class DisplacementGrid}
}
\value{
a sitk::Image<sitkVectorFloat32>
}
\description{
convert an object of class DisplacementGrid to deformation field created by mesheR::displacementField2Grid
}

