% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill3DVectorImage.r
\name{fillVectorImage}
\alias{fillVectorImage}
\title{this a first test whether it is possible to call SimpleITK via Rcpp}
\usage{
fillVectorImage(image, indices, values)
}
\arguments{
\item{image}{image of class}

\item{vector}{of length}
}
\value{
returns the image dimension
}
\description{
this a first test whether it is possible to call SimpleITK via Rcpp. It simply shows how to change the origin of a sitk::Image using SimpleITK's C++ interface.
}
\examples{
require(SimpleITK)
img <- Image(10,10,10,"sitkVectorFloat32")
indices <- as.matrix(expand.grid(1:10,1:10,1:10))
set.seed(42)
values <- matrix(rnorm(3000),1000,3)
img <- fillVectorImage(img,indices,values)
stopifnot(all.equal(img[1,1,1],values[1,]))
}

