% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{GetImageDimension}
\alias{GetImageDimension}
\title{this a first test whether it is possible to call SimpleITK via Rcpp}
\usage{
GetImageDimension(image)
}
\arguments{
\item{image}{image of class}
}
\value{
returns the image dimension
}
\description{
this a first test whether it is possible to call SimpleITK via Rcpp
}
\examples{
require(SimpleITK)
img <- Image(10,10,10,"sitkInt8")
print(GetImageDimension(img))
}

