% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillVectorImage.r
\name{fillVectorImage}
\alias{fillVectorImage}
\title{this a first test whether it is possible to call SimpleITK via Rcpp}
\usage{
fillVectorImage(image, indices, values, float = FALSE)
}
\arguments{
\item{image}{SimpleITK image}

\item{indices}{integer matrix with rows containing (1-based) indices}

\item{values}{numeric matrix with rows containing vectors}

\item{float}{if TRUE, vectors will be coerced to float (instead of double)}
}
\value{
returns the updated image
}
\description{
this a first test whether it is possible to call SimpleITK via Rcpp. It simply shows how to change the origin of a sitk::Image using SimpleITK's C++ interface.
}
\examples{
require(SimpleITK)
img <- Image(10,10,10,"sitkVectorFloat32")
indices <- as.matrix(expand.grid(1:10,1:10,1:10))
set.seed(42)
values <- matrix(rnorm(3000),1000,3)
img <- fillVectorImage(img,indices,values)
stopifnot(all.equal(img[1,1,1],values[1,]))
}

